Refactor functions into oxidationstates.py and informatics.py
